requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
firebase-admin==6.5.0
google-generativeai>=0.8.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyMuPDF>=1.26.4
numpy>=1.25.0
tqdm>=4.66.1
psutil>=5.9.0
